separate donation model from view
