+++ lib/donations.dart
@@ -1,0 +1,6 @@
+// model for a list of donations
+import 'package:ovcapp/donation.dart';
+
+class Donations {
+  List<Donation> donations = [];
+}
